♻️ refactor: client 접근 cookie와 server 접근 cookie 분리
