+++ src/utils/index.ts
@@ -1,2 +1,3 @@
-export * from './cookie';
+export * from './cookie.server';
+export * from './cookie.client';
 export * from './isServer';
